Use BlockDraggable instead of Draggable
This makes BlockDraggable DOM agnostic and uses it as a base to
compose drag data. The vision is that we may have an ImageDraggable
component in the future that would abstract what data the dragging
operation needs to drag images, and so on.

The goal would be to use it by only providing one prop, the
blockClientId.

Making this a single commit in case we want to revert.